test: improve test cases for svelte plugin
