+++ e2e/cases/plugin-less/exclude-with-helper/rsbuild.config.ts
@@ -1,6 +1,15 @@
+import { defineConfig } from '@rsbuild/core';
 import { pluginLess } from '@rsbuild/plugin-less';
 
-export default {
+export default defineConfig({
+  tools: {
+    bundlerChain(chain) {
+      chain.module
+        .rule('fallback')
+        .test(/b\.less$/)
+        .type('asset/resource');
+    },
+  },
   plugins: [
     pluginLess({
       lessLoaderOptions: (_, { addExcludes }) => {
@@ -8,4 +17,4 @@ export default {
       },
     }),
   ],
-};
+});
