+++ e2e/cases/plugin-less/exclude/index.test.ts
@@ -3,19 +3,7 @@ import { expect, test } from '@e2e/helper';
 test('should exclude specified Less files using the exclude option', async ({
   build,
 }) => {
-  const rsbuild = await build({
-    rsbuildConfig: {
-      tools: {
-        bundlerChain(chain) {
-          chain.module
-            .rule('fallback')
-            .test(/b\.less$/)
-            .type('asset/resource');
-        },
-      },
-    },
-  });
-
+  const rsbuild = await build();
   const files = rsbuild.getDistFiles();
   const cssFiles = Object.keys(files).filter((file) => file.endsWith('.css'));
   const lessFiles = Object.keys(files).filter((file) => file.endsWith('.less'));
