+++ e2e/cases/plugin-less/exclude/rsbuild.config.ts
@@ -1,9 +1,18 @@
+import { defineConfig } from '@rsbuild/core';
 import { pluginLess } from '@rsbuild/plugin-less';
 
-export default {
+export default defineConfig({
   plugins: [
     pluginLess({
       exclude: /b\.less$/,
     }),
   ],
-};
+  tools: {
+    bundlerChain(chain) {
+      chain.module
+        .rule('fallback')
+        .test(/b\.less$/)
+        .type('asset/resource');
+    },
+  },
+});
