+++ e2e/cases/plugin-sass/exclude-with-helper/index.test.ts
@@ -3,19 +3,7 @@ import { expect, test } from '@e2e/helper';
 test('should exclude specified Sass files using addExcludes', async ({
   build,
 }) => {
-  const rsbuild = await build({
-    rsbuildConfig: {
-      tools: {
-        bundlerChain(chain) {
-          chain.module
-            .rule('fallback')
-            .test(/b\.scss$/)
-            .type('asset/resource');
-        },
-      },
-    },
-  });
-
+  const rsbuild = await build();
   const files = rsbuild.getDistFiles();
   const cssFiles = Object.keys(files).filter((file) => file.endsWith('.css'));
   const scssFiles = Object.keys(files).filter((file) => file.endsWith('.scss'));
