+++ e2e/cases/plugin-sass/exclude-with-helper/rsbuild.config.ts
@@ -1,6 +1,15 @@
+import { defineConfig } from '@rsbuild/core';
 import { pluginSass } from '@rsbuild/plugin-sass';
 
-export default {
+export default defineConfig({
+  tools: {
+    bundlerChain(chain) {
+      chain.module
+        .rule('fallback')
+        .test(/b\.scss$/)
+        .type('asset/resource');
+    },
+  },
   plugins: [
     pluginSass({
       sassLoaderOptions: (_, { addExcludes }) => {
@@ -8,4 +17,4 @@ export default {
       },
     }),
   ],
-};
+});
