+++ e2e/cases/web-worker/web-worker-target/index.test.ts
@@ -1,5 +1,4 @@
 import path from 'node:path';
-
 import { expect, test } from '@e2e/helper';
 
 test('should build web-worker target correctly', async ({ build }) => {
